Appiled Cryptography/CA2/requirements.txt
pyftpdlib>=1.5.6
pycryptodome
pyopenssl
pillow